# Optimize pipeline task run query