--- a/core/services/pipeline/orm.go
+++ b/core/services/pipeline/orm.go
@@ -186,20 +186,15 @@ func (o *orm) CreateRun(ctx context.Context, jobID int32, meta map[string]interf
 		runID = run.ID
 
 		// Create the task runs
-		err = tx.Exec(`
+		var trs []TaskRun
+		err = tx.Raw(`
             INSERT INTO pipeline_task_runs (
             	pipeline_run_id, pipeline_task_spec_id, type, index, created_at, predecessor_task_run_ids
             )
             SELECT ? AS pipeline_run_id, id AS pipeline_task_spec_id, type, index, NOW() AS created_at, '{}'
             FROM pipeline_task_specs
-            WHERE pipeline_spec_id = ?`, run.ID, run.PipelineSpecID).Error
-		if err != nil {
-			return err
-		}
-
-		// Ammend the TaskRuns with their predecessors.
-		var trs []TaskRun
-		err = tx.Find(&trs).Where("pipeline_run_id = ?", runID).Error
+            WHERE pipeline_spec_id = ?
+			RETURNING *`, run.ID, run.PipelineSpecID).Scan(&trs).Error
 		if err != nil {
 			return err
 		}