--- a/core/store/migrations/migrate.go
+++ b/core/store/migrations/migrate.go
@@ -3,6 +3,8 @@ package migrations
 import (
 	"regexp"
 
+	"github.com/smartcontractkit/chainlink/core/store/migrations/migration1611174407"
+
 	"github.com/smartcontractkit/chainlink/core/store/migrations/migration1609963213"
 
 	"github.com/smartcontractkit/chainlink/core/store/migrations/migration1608217193"
@@ -492,6 +494,10 @@ func init() {
 		{
 			ID:      "1610630629",
 			Migrate: migration1610630629.Migrate,
+		},
+		{
+			ID:      "1611174407",
+			Migrate: migration1611174407.Migrate,
 		},
 	}
 }