--- a/go.mod
+++ b/go.mod
@@ -60,5 +60,6 @@ require (
 	gonum.org/v1/gonum v0.8.2
 	gopkg.in/gormigrate.v1 v1.6.0
 	gopkg.in/guregu/null.v4 v4.0.0
+	gorm.io/driver/postgres v1.0.6
 	gorm.io/gorm v1.20.11
 )