At the scale of the load test (15k task runs) the speed up is (1.859594234s/8.343933ms) = ~222x. Not totally sure why the local explain analyze is waay slower for the old query vs running the load test but here are the two explains:

*Before*
```
| Limit  (cost=1646.30..1647.01 rows=1 width=130) (actual time=70185.760..70185.925 rows=1 loops=1)
|   ->  Nested Loop  (cost=1646.30..3425.83 rows=2513 width=130) (actual time=70185.747..70185.895 rows=1 loops=1)
|         ->  HashAggregate  (cost=1646.01..1708.51 rows=2513 width=8) (actual time=70185.687..70185.811 rows=1 loops=1)
|               Group Key: pipeline_task_runs_1.id
|               Filter: (bool_and((predecessor_unfinished_runs.finished_at IS NOT NULL)) OR (count(predecessor_unfinished_runs.id) = 0))
|               ->  Hash Right Join  (cost=416.24..1441.47 rows=27273 width=24) (actual time=139.404..34446.488 rows=5000000 loops=1)
|                     Hash Cond: ((predecessor_unfinished_runs.pipeline_task_spec_id = predecessor_specs.id) AND (predecessor_unfinished_
|                     ->  Seq Scan on pipeline_task_runs predecessor_unfinished_runs  (cost=0.00..340.00 rows=15000 width=28) (actual tim
|                     ->  Hash  (cost=341.24..341.24 rows=5000 width=20) (actual time=139.314..139.403 rows=5000 loops=1)
|                           Buckets: 8192  Batches: 1  Memory Usage: 318kB
|                           ->  Hash Join  (cost=46.49..341.24 rows=5000 width=20) (actual time=0.736..104.121 rows=5000 loops=1)
|                                 Hash Cond: (pipeline_task_runs_1.pipeline_task_spec_id = pipeline_task_specs.id)
|                                 ->  Index Scan using idx_pipeline_task_runs_unfinished on pipeline_task_runs pipeline_task_runs_1  (cos
|                                 ->  Hash  (cost=39.33..39.33 rows=550 width=8) (actual time=0.678..0.732 rows=15 loops=1)
|                                       Buckets: 1024  Batches: 1  Memory Usage: 9kB
|                                       ->  Hash Right Join  (cost=22.38..39.33 rows=550 width=8) (actual time=0.315..0.612 rows=15 loops
|                                             Hash Cond: (predecessor_specs.successor_id = pipeline_task_specs.id)
|                                             ->  Seq Scan on pipeline_task_specs predecessor_specs  (cost=0.00..15.50 rows=550 width=8)
|                                             ->  Hash  (cost=15.50..15.50 rows=550 width=4) (actual time=0.266..0.284 rows=15 loops=1)
|                                                   Buckets: 1024  Batches: 1  Memory Usage: 9kB
|                                                   ->  Seq Scan on pipeline_task_specs  (cost=0.00..15.50 rows=550 width=4) (actual time
|         ->  Index Scan using pipeline_task_runs_pkey on pipeline_task_runs  (cost=0.29..0.67 rows=1 width=130) (actual time=0.039..0.04
|               Index Cond: (id = pipeline_task_runs_1.id)
| Planning Time: 0.533 ms
| Execution Time: 70186.177 ms
```

*After*
```
 | QUERY PLAN
|----------------------------------------------------------------------------------------------------------------------------------------
| Limit  (cost=0.86..8.12 rows=1 width=130) (actual time=0.273..0.382 rows=1 loops=1)
|   ->  Merge Join  (cost=0.86..18269.00 rows=2513 width=130) (actual time=0.256..0.342 rows=1 loops=1)
|         Merge Cond: (pipeline_task_runs.id = runs.id)
|         ->  Index Scan using pipeline_task_runs_pkey on pipeline_task_runs  (cost=0.29..590.28 rows=15000 width=130) (actual time=0.022
|         ->  GroupAggregate  (cost=0.57..17584.67 rows=2513 width=8) (actual time=0.175..0.229 rows=1 loops=1)
|               Group Key: runs.id
|               Filter: (bool_and((preds.finished_at IS NOT NULL)) OR (count(preds.*) = 0))
|               ->  Nested Loop Left Join  (cost=0.57..17147.28 rows=49985 width=170) (actual time=0.082..0.185 rows=2 loops=1)
|                     ->  Index Scan using pipeline_task_runs_pkey on pipeline_task_runs runs  (cost=0.29..590.28 rows=5000 width=33) (ac
|                           Filter: (finished_at IS NULL)
|                           Rows Removed by Filter: 4
|                     ->  Index Scan using pipeline_task_runs_pkey on pipeline_task_runs preds  (cost=0.29..3.21 rows=10 width=170) (actu
|                           Index Cond: (id = ANY (runs.predecessor_task_run_ids))
| Planning Time: 0.906 ms
| Execution Time: 0.576 ms
```